+++ packages/mock-app/src/pages/Verify.tsx
@@ -1,5 +1,5 @@
 import { Status } from '@mock-app/components';
-import { computeHash, decryptCredential, privateAPI, publicAPI, verifyVC } from '@mock-app/services';
+import { computeHash, decryptCredential, publicAPI, verifyVC } from '@mock-app/services';
 import { IVerifyResult, VerifiableCredential } from '@vckit/core-types';
 import * as jose from 'jose';
 import React, { useCallback, useEffect, useState } from 'react';
